chore: I18n extract compile (#5540)
## Description

This PR runs extract and compile on the i18n translations.

## Notes

The reason this came up is because an upcoming change in a separate PR
requires a copy change to a message in English that is already in the
compiled catalog (and so will require either a spot change in that
particular catalog or that this pipeline is run)

To avoid having to manually change the catalog, or create a large diff
in the other PR, I've put this PR up that only runs the extract/compile
pipeline (which has not been run for some time).