[Vis Tools] Implement redirect of "new" visualization tools to "old" tools (#5502)
This PR implements a redirect from
`/tools/visualization#visType=timeline to the equivalent
/tools/timeline#` page when the `standardized_vis_tool` flag is enabled.

More specifically:
* Query parameters are now propagated on /tools/visualization pages.
This is to ensure `?disable_feature` works in tests.

* The tests for /tools/visualization have been updated to explicitly
disable the `standardized_vis_tool` flag

* A new `redirect_utils.ts` file is added with the functions used to
translate a /tools/visualization url to a /tools/timeline url

* The test for the /tools/visualization timeline-specific landing page
links has been removed, because those links now result in a redirect.

## Testing Strategy

1. Checkout these changes and pin up a local instance of the website
repo

1. Go to `http://localhost:8080/tools/visualization#visType=timeline`
and see that you are redirected to the old timeline tool at
/tools/timeline.

1. Go to the original visualization tool timeline landing page by going
to `http://localhost:8080/tools/visualization`

1. Click on the 'Timeline' tab

1. Click through each of the links on the page, see that they link to a
corresponding /tool/timeline page, and that each of the charts renders
correctly.