Make a mixer-only version of full-env image (#5567)
An image with all of go, yq, curl, and kubectl is needed to validate and
deploy mixer feature flags on Cloud Build. The same image can also be
used for pre-merge checks in the mixer repo.

- Excluding Chrome, Node, and Envoy cuts the image size in half
- Adding golangci-lint reduces the number of images needed to run
pre-merge checks

Since this image has the right version of Python and is capable of
running webdriver tests, it can also be used in other webdriver build
trigger steps.