Reduce website staging V3 mirror fraction (#5574)
This value got bumped when we switched to using Cloud Deploy, since we
went from using Helm charts in website repo to using the ones in mixer
repo with overrides defined in skaffold.yaml.

- Before Cloud Deploy:
https://github.com/datacommonsorg/website/blob/afea0ba1dc70d1680146a4cbafd04bfc6f26a062/deploy/helm_charts/envs/staging.yaml#L48
- After Cloud Deploy:
https://github.com/datacommonsorg/website/blob/afea0ba1dc70d1680146a4cbafd04bfc6f26a062/skaffold.yaml#L321

In the medium term, we should also resolve other discrepancies between
the website repo vs mixer repo per-environment values, such as number of
replicas per serviceGroup. But prioritizing this quick fix in the short
term to stop the unexpected increase in our Spanner usage.