[Stat Var Autocomplete] Improve Google Analytics collection for Experiment (#5585)
Move the logic to decide whether to render the SV autocomplete into the
client so we can properly log GA events with that breakdown.

Force the server to respect the FE's decision (for experiments) by using
the override flags to enable/disable a feature.

This allows to log the number of times the NL search bar is rendered
broken down by this experiment + the number of explore_search_q events
by this experiment. This will allow us to accurately identify if
enabling this experiment has an outsized impact on total queries
executed.