[Vis Tools] Redirect from visualization tools to /tools/map (#5583)
This PR implements a redirect from /tools/visualization#visType=map to
the equivalent /tools/map# page when the standardized_vis_tool flag is
enabled.

In particular:

* The redirect check for /tools/visualization has been simplified now
that redirects are enabled for all tools
* Redirecting when clicking on a tab has been simplified now that
redirects are enabled for all tools
* The "Switch tool" button has been removed for /tools/map

## Testing Strategy

1. Checkout these changes and spin up a local instance of the website
repo

1. Go to http://localhost:8080/tools/visualization. See that you are
redirected to the old map tool at /tools/map.

1. Go to http://localhost:8080/tools/visualization#visType=map and see
that you are redirected to the old scatter tool at /tools/map.

1. Go to the original visualization tool landing page by going to
http://localhost:8080/tools/visualization?disable_feature=standardized_vis_tool

1. Click through each of the links on the page. If you remove
disable_feature=standardized_vis_tool from the URL, see that they link
to a corresponding /tool/map page, and that each of the charts renders
correctly.