API Dialog - Python and Syntax Highlighting (#5579)
## Description

This PR implements the following features in the API dialog:
* Syntax highlighting
* Python endpoints

## Libraries

To effect the highlighting itself, we used Prism (prism.js), as it is
lightweight, extensible and very well used.

## Notes

To allow for multiple languages, we introduced a "language spec", object
that describe a given language and indicate the conversion function that
will be used to convert an observation spec into the API call in that
language.

Some languages and protocols such as Python are well suited to single
scripts that can contain multiple API calls. Others, such as cURL, need
to have each call separated. Whether a particular language is "single
script" or "multiple endpoint" is indicated by that languages spec. The
language will have an appropriate generator function based on what type
it is.

Instructions are provided in the comments of the API dialog component
that describe how a new language can be added.

### Scrolling

We had discussed the idea of having the scrolling within the elements
that contain the code (to keep the height of the dialog stable).
However, this results in double-scrolling when the dialog is small (and
is also inconsistent because of the fact that some languages will
contain more than one code section.

Instead, we allowed the dialog to handle the scrolling as per normal,
and put in a fade transition to help smooth the effect.

## Screenshots

<img width="1018" height="732" alt="Screenshot 2025-10-04 at 11 44
48 AM"
src="https://github.com/user-attachments/assets/5740a8be-a5e5-47c3-9e03-944815ed62e4"
/>

<img width="989" height="536" alt="Screenshot 2025-10-02 at 8 17 13 PM"
src="https://github.com/user-attachments/assets/8ce85905-2d5a-4feb-af01-3eea8c18c8c1"
/>

## Tests

The wedriver test for the API dialog has been updated for changes made
by this PR, as well as augmented to test language switching via the
select box.

---------

Co-authored-by: Pablo Noel <pablo@next-solutions.ca>