Download dialog - CSV highlighting (#5654)
## Description

During the testing of the download tool updates found in a [PR
5653](https://github.com/datacommonsorg/website/pull/5653), it became
noticeable that the data preview window in the Download dialog was often
difficult to read, enough so that I downloading the CSVs rather than
looking at the data directly in the preview.

To improve the readability, this PR introduces the CSV language to our
code block, and converts that display to the code block.

Besides the updated coloring, we now display each line on its own (with
horizontal scrolling) to make it easier for the user to scan.

## Notes

In the long term, we could consider converting this window into a little
data table (designing the data table in a way that it could also be used
in other places, such as in the preview associated with the general
download tool that will be redone soon).

## Screenshots

### Before

<img width="977" height="619" alt="before"
src="https://github.com/user-attachments/assets/e8184342-c2a4-44a2-9da5-6fe805149c23"
/>

### After

<img width="1005" height="631" alt="Screenshot 2025-10-26 at 4 42 29 PM"
src="https://github.com/user-attachments/assets/8ba90aa2-824d-4226-b98d-8835d91fd542"
/>

---------

Co-authored-by: Pablo Noel <pablo@next-solutions.ca>