Updating nodejs/query surface handling to parse text correctly (#5673)
This initially handled the case that the header was an array incorrectly
and only returned the first character of the header if it was a string.
This fixes that so the entire surface is returned. We don't expect the
surface to be a list but this handles it gracefully if it somehow is.
This is also safe if the surface is undefined, in which case it gets set
to "unknown" when passed into mixer.

This can be tested by running the nodejs server locally following the
instructions
[here](https://github.com/datacommonsorg/website/tree/master/static/src#start-server).
I ran the website with `./run_server.sh -m -l` to use local mixer, then
manually verified that the surface in my usage logs is now correct.