Update Node image; Try to speed up tests (#5571)
- Make a fresh version of the Docker image with pre-built node_modules,
including the two other directories that have them.
  - Add convenience script for copying them into place
- Also disable npm audit when using them, on the suspicion that npm
downloads a bunch of stuff despite the cache to do the audit.

~Fold in the Custom DC webdriver tests that used to run single-threaded
under the `one_at_a_time` annotation. Recent efforts to reduce flakiness
hopefully mean that this is safe. If that proves out after a couple
weeks, we can go through and clean up `one_at_a_time` elsewhere in the
code.~ Left this out in the end because it didn't seem to improve
flakiness.