[Vis Tools] Fix Scatter Tool context handling to address flickering and race conditions (#5727)
# Changes Made
This PR fixes some bugs in the Scatter Tool related to context
initialization and updating. These changes are:

### 1. Initialize Context using hash parameter values

Currently, the Scatter tool initializes its context with an empty state,
and requires a re-render when updating the empty state with hash values
from the URL.

This PR initializes the context with values from the current hash, which
prevents the unnecessary re-render.

### 2. Use React's builtin previous state handling to avoid race
conditions

The current context setters had the client manually provide the previous
state to update. The problem is that this resulted in a race condition
when multiple context setting functions are called concurrently. Changes
in one context update would get clobbered by a different context update
because the second updater was using a now-outdated version of the
context.

This PR updates the setters to use react's built-in functional state
update syntax so that react always sees the latest `prevState`.

### 3. Clean up handling of hash parameters in utils
The context handling utils contain a bug that result in hash parameters
being added even when the parameter's value is undefined. This PR now
only adds hash parameters to the URL if both key and value are truthy.

### 4. Update client tests
This PR also updates the scatter npm client tests to match these
changes.

# Issues Fixed
As a result, these changes fix the following issues:

### 1. The stat var widget flickers in and out when loading a chart URL
When clicking landing page links or directly navigating to a scatter
chart URL, the stat var widget will flicker in and out due to the
enclosing places being re-set between re-renders. The functional version
of state updates does not re-set enclosing places between re-renders,
which stops the flickering.

Before:


https://github.com/user-attachments/assets/05abbe2b-3cde-435b-94d9-779054022f21


After:


https://github.com/user-attachments/assets/60e83747-9115-4bdf-8acd-c9f7cd0095a1


### 2. The scatter tool hangs sometimes
Sometimes, the tool will stall and not load anything despite having
selected settings via the UI. This is caused by the race condition
described above, where concurrent context updates clobber each other,
leaving the app in an invalid state. Switching to the functional state
update syntax removes this race condition.

### 3. Scatter tool URLs have unnecessary entries
Interacting with the scatter tool results in chart URLs that look like
this (note the unnecessary `%26` after the hash `#`):
```
/tools/scatter#%26svx%3DCount_Person%26lx%3D1%26dx%3DCount_Person%26svy%3DCount_HousingUnit%26pcy%3D1%26dy%3DCount_Person%26epd%3DgeoId%2F10%26ept%3DCounty%26ub%3D99999%26qd%3D1%26ld%3D1%26dd%3D1%26rg%3D1%26pp%3Dlinear
```
Note the unnecessary %26 in front. With this change, URLs now look like
(note the unnecessary `%26` after the hash `#` is now gone):
```
/tools/scatter#svx%3DCount_Person%26lx%3D1%26dx%3DCount_Person%26svy%3DCount_HousingUnit%26pcy%3D1%26dy%3DCount_Person%26epd%3DgeoId%2F10%26ept%3DCounty%26ub%3D99999%26qd%3D1%26ld%3D1%26dd%3D1%26rg%3D1%26pp%3Dlinear
```