Adds scripts to automate usage log cloud setup (#5505)
- Split into two scripts, one that creates a dataset and cloud loggers
in each project that has a Mixer instance and one that creates a
aggregation dataset and schedules a query to aggregate the logs.
- I split these up because Cloud Logging takes a minute to ingest the
first batch of logs and declare the schema of the destination table.
It's possible to manually define the schema, but because it's so complex
it seems easier especially if partners modify the schema in the future
to just wait.
- Also includes a readme outlining how to use the scripts.
- Fully parameterized so can be customized for any GCP project.