Metadata and download modal - Date display (line charts) (#5679)
## Description

This PR implements the ability for the metadata modal and chart embed
dialog (the download dialog) to display the date range displayed in the
actual chart.

It can be divided into two distinct parts:
* The update of the metadata modal and citations display to include the
actual chart data range (details below).
* An update to the line tile to feed the date range out from the chart
and up to where the metadata modal and the embed dialog have access to
it. This parallels the exposure of other data such as the stat var and
facet details that the chart already delivers.

To keep the size of the PR down, it implements the chart range only for
the line charts. Subsequent charts (map, scatter, highlight, bar,
ranking) will follow in a coming PR.

## Display Logic

In the metadata modal, we use the following logic (this is also included
as a comment in the code):
* If we have neither a date range from the metadata nor the chart, we
display nothing.
* If we have only a metadata date range, we display that.
* If we have only a chart-derived date range, we display that.
* If we have both but they are the same, we display that single shared
date range.
* If we have both and they are different, we display both with the chart
range in parentheses (see screenshot).

In the citation, if we have a chart-derived date range, we display it.
This applies to both the metadata modal and the embed chart dialog.

## Testing

For line charts, many facets have the same date range displayed in the
chart as stored in the metadata. This means that for most facets, the
only visible difference will be that the chart-derived date range
appears in the citation.

For line charts where the metadata does not display a date, we now show
the chart derived date.

For line charts where the chart-derived date is different, we display
both. This is rare for line charts (but will be more common with charts
like map charts). However, it can be found in the wikipedia data
example. To replicate this while testing, temporarily comment out the
wikipedia entry in the suppression list in `metadata_fetcher.ts`:

```
const DATE_RANGE_SUPPRESSION_PROVENANCES = [];
//const DATE_RANGE_SUPPRESSION_PROVENANCES = ["Wikidata", "WikipediaStatsData"];
```

* Visit this page: [Test
search](http://localhost:8080/explore?enable_feature=enable_stat_var_autocomplete#q=population+united+states&client=ui_query_)
* With the default facet (PEP), open the "About this data". You will see
the date range 1900-2024 in the metadata (this is both the chart-derived
and metadata derived range) and 1900-2024 in the citation (this is new).
* Open the Download dialog. You will see the same date range in the
citation footer.
* Choose Wikidata as the facet. If you have removed that entry from the
suppression list, you will see both the metadata-derived date and the
chart-derived date.
* If you keep Wikidata in the suppression list (i.e., what will happen
when this is merged to master), you will see the chart-derived data in
the metadata date range (there is no metadata-derived range available)

## Screenshots

### Metadata and Chart-derived dates are the same
_Difference from before: the date is displaying in the citation footer_

<img width="744" height="725" alt="Metadata and chart derived data the
same"
src="https://github.com/user-attachments/assets/6cf6cc67-653b-422c-bdb7-e9f65dac254f"
/>

### No metadata range for facet
_Difference from before: Before we showed no date at all. Now We now
show the chart-derived range in the stat var details (as well as the
citation footer)._

<img width="745" height="586"
alt="Wikipedia-no-metadata-range-available"
src="https://github.com/user-attachments/assets/4447ab02-7bc6-4eb0-8b6c-b0635f94f527"
/>

### Metadata range for the facet is different than the chart-derived
range
_Difference from before: before we would just show the metadata range.
Now we show the actual range in parentheses next to it, and also show
the actual range in the citation footer._

<img width="751" height="585" alt="Screenshot 2025-11-02 at 9 02 16 PM"
src="https://github.com/user-attachments/assets/9db52c17-09da-495b-8d10-5bedd1ccf92e"
/>

### Chart embed (Download) dialog
_Difference from before: We now display the chart-derived date here._

<img width="983" height="625" alt="chart_download"
src="https://github.com/user-attachments/assets/4697c069-b804-47ae-8020-8c896327f49d"
/>