Update tsconfig.json target to "es2021" and module to "esnext" (#5839)
As part of #5835 , the target in tsconfig.json needs to be upgraded from
es6 to es2021. This PR isolates the es2021 change.

The main difference appears to be how dynamic imports are treated. This
PR:
* Updates the module in tsconfig.json to "esnext" instead of "commonJS".
This allows webpack to correctly use `import()` as we have across the
codebase
* Updates the handling of dynamic imports (primarily in
`loadLocaleData()`) to be es2021 compatible
* Includes some trailing lint fixes to get pre-merge tests to pass

---------

Co-authored-by: Xiao Shi <shixiao@google.com>